[Elasticsearch] Move PaginatedArray to own file
